second

+++ UCTNodePointer.cpp
@@ -114,10 +114,10 @@ bool UCTNodePointer::valid() const {
     return true;
 }
 
-int UCTNodePointer::get_visits() const {
+double UCTNodePointer::get_visits() const {
     auto v = m_data.load();
     if (is_inflated(v)) return read_ptr(v)->get_visits();
-    return 0;
+    return 0.0;
 }
 
 float UCTNodePointer::get_policy() const {
